requirements.txt
Django==5.1.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-filter==24.2
drf-yasg==1.21.8

django-cors-headers==4.3.1
psycopg2-binary
python-dotenv==1.0.1

# Ratelimit (to'g'ri versiya)
django-ratelimit==4.1.0

Pillow==10.4.0
